+++ libraries/libwidget/widgets/ScrollBar.cpp
@@ -32,12 +32,12 @@ void ScrollBar::scroll_to(Vec2i position)
     if (_horizontal)
     {
         int new_value = position.x() - track_bound().x();
-        value((new_value / (double)track_bound().width()) * _track);
+        value((new_value / (float)track_bound().width()) * _track);
     }
     else
     {
         int new_value = position.y() - track_bound().y();
-        value((new_value / (double)track_bound().height()) * _track);
+        value((new_value / (float)track_bound().height()) * _track);
     }
 
     Event event_value_changed = {};
